move converter factory to a new file
