+++ experimental/components/tests/units/units_test.rs
@@ -4,7 +4,7 @@
 
 use core::str::FromStr;
 
-use icu_experimental::units::converter::ConverterFactory;
+use icu_experimental::units::converter_factory::ConverterFactory;
 use num_bigint::BigInt;
 use num_rational::BigRational;
 use num_rational::Ratio;
